pyproject.toml
[tool.poetry]
name = "exercicio4"
version = "0.1.0"
description = ""
authors = ["Marcelo Schneider <schenider.fei@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyspark = "^3.4.1"
pytest = "^7.4.0"
ftfy = "^6.1.1"
chardet = "^5.2.0"
python-decouple = "^3.8"
unidecode = "^1.3.6"
pandas = "^2.1.0"
pyarrow = "^13.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
